requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==0.28.1
pandas==2.2.0
openpyxl==3.1.2
apscheduler==3.10.4
thefuzz==0.20.0
transliterate==1.10.2
alembic==1.13.0